Feature/CORE-130/21 async mempool database

Approved-by: Arkadiusz Kolodziejski
Approved-by: Richard Mills

+++ src/test/merkle_tests.cpp
@@ -147,7 +147,7 @@ BOOST_AUTO_TEST_CASE(merkle_test) {
                 for (int loop = 0; loop < std::min(ntx, 16); loop++) {
                     // If ntx <= 16, try all branches. Otherise, try 16 random
                     // ones.
-                    int mtx = loop;
+                    size_t mtx = loop;
                     if (ntx > 16) {
                         mtx = InsecureRandRange(ntx);
                     }
